+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix an issue with `linkTo` which would use legacy verb `navigate` instead of `set page`
 - `render` can now take properties that have deeply nested generators
 - Added `history` and `pop` streams to stdlib. Basic popstate functionality.
+- Added option to use URL fragments (i.e. "hash URLs"), just set `window.root` to the folder where index files are located (can be relative)
 
 # 0.4.1